+++ tickMeter/Forms/SettingsForm.Designer.cs
@@ -63,17 +63,8 @@ namespace tickMeter.Forms
             this.run_minimized = new System.Windows.Forms.CheckBox();
             this.local_ip_lbl = new System.Windows.Forms.Label();
             this.local_ip_textbox = new System.Windows.Forms.TextBox();
-            this.captureAllAdaptersCheckbox = new System.Windows.Forms.CheckBox();
-            this.chkPingBindToInterface = new System.Windows.Forms.CheckBox();
-            this.chkPingTcpPrefer = new System.Windows.Forms.CheckBox();
             this.chkPingFallbackIcmp = new System.Windows.Forms.CheckBox();
-            this.chkPingTargetActiveOnly = new System.Windows.Forms.CheckBox();
-            this.chkTickrateSmoothing = new System.Windows.Forms.CheckBox();
-            this.chkDedupMultiNic = new System.Windows.Forms.CheckBox();
-            this.chkEnableIPv6 = new System.Windows.Forms.CheckBox();
             this.chkIgnoreVirtualAdapters = new System.Windows.Forms.CheckBox();
-            this.chkRtssOnlyActive = new System.Windows.Forms.CheckBox();
-            this.chkStunEnable = new System.Windows.Forms.CheckBox();
             this.label8 = new System.Windows.Forms.Label();
             this.donate_lbl = new System.Windows.Forms.Label();
             this.updateLbl = new System.Windows.Forms.Label();
@@ -357,27 +348,6 @@ namespace tickMeter.Forms
             this.local_ip_textbox.Name = "local_ip_textbox";
             this.local_ip_textbox.TextChanged += new System.EventHandler(this.local_ip_textbox_TextChanged);
             // 
-            // captureAllAdaptersCheckbox
-            // 
-            resources.ApplyResources(this.captureAllAdaptersCheckbox, "captureAllAdaptersCheckbox");
-            this.captureAllAdaptersCheckbox.ForeColor = System.Drawing.Color.Black;
-            this.captureAllAdaptersCheckbox.Name = "captureAllAdaptersCheckbox";
-            this.captureAllAdaptersCheckbox.UseVisualStyleBackColor = true;
-            // 
-            // chkPingBindToInterface
-            // 
-            resources.ApplyResources(this.chkPingBindToInterface, "chkPingBindToInterface");
-            this.chkPingBindToInterface.ForeColor = System.Drawing.Color.Black;
-            this.chkPingBindToInterface.Name = "chkPingBindToInterface";
-            this.chkPingBindToInterface.UseVisualStyleBackColor = true;
-            // 
-            // chkPingTcpPrefer
-            // 
-            resources.ApplyResources(this.chkPingTcpPrefer, "chkPingTcpPrefer");
-            this.chkPingTcpPrefer.ForeColor = System.Drawing.Color.Black;
-            this.chkPingTcpPrefer.Name = "chkPingTcpPrefer";
-            this.chkPingTcpPrefer.UseVisualStyleBackColor = true;
-            // 
             // chkPingFallbackIcmp
             // 
             resources.ApplyResources(this.chkPingFallbackIcmp, "chkPingFallbackIcmp");
@@ -385,54 +355,12 @@ namespace tickMeter.Forms
             this.chkPingFallbackIcmp.Name = "chkPingFallbackIcmp";
             this.chkPingFallbackIcmp.UseVisualStyleBackColor = true;
             // 
-            // chkPingTargetActiveOnly
-            // 
-            resources.ApplyResources(this.chkPingTargetActiveOnly, "chkPingTargetActiveOnly");
-            this.chkPingTargetActiveOnly.ForeColor = System.Drawing.Color.Black;
-            this.chkPingTargetActiveOnly.Name = "chkPingTargetActiveOnly";
-            this.chkPingTargetActiveOnly.UseVisualStyleBackColor = true;
-            // 
-            // chkTickrateSmoothing
-            // 
-            resources.ApplyResources(this.chkTickrateSmoothing, "chkTickrateSmoothing");
-            this.chkTickrateSmoothing.ForeColor = System.Drawing.Color.Black;
-            this.chkTickrateSmoothing.Name = "chkTickrateSmoothing";
-            this.chkTickrateSmoothing.UseVisualStyleBackColor = true;
-            // 
-            // chkDedupMultiNic
-            // 
-            resources.ApplyResources(this.chkDedupMultiNic, "chkDedupMultiNic");
-            this.chkDedupMultiNic.ForeColor = System.Drawing.Color.Black;
-            this.chkDedupMultiNic.Name = "chkDedupMultiNic";
-            this.chkDedupMultiNic.UseVisualStyleBackColor = true;
-            // 
-            // chkEnableIPv6
-            // 
-            resources.ApplyResources(this.chkEnableIPv6, "chkEnableIPv6");
-            this.chkEnableIPv6.ForeColor = System.Drawing.Color.Black;
-            this.chkEnableIPv6.Name = "chkEnableIPv6";
-            this.chkEnableIPv6.UseVisualStyleBackColor = true;
-            // 
             // chkIgnoreVirtualAdapters
             // 
             resources.ApplyResources(this.chkIgnoreVirtualAdapters, "chkIgnoreVirtualAdapters");
             this.chkIgnoreVirtualAdapters.ForeColor = System.Drawing.Color.Black;
             this.chkIgnoreVirtualAdapters.Name = "chkIgnoreVirtualAdapters";
             this.chkIgnoreVirtualAdapters.UseVisualStyleBackColor = true;
-            // 
-            // chkRtssOnlyActive
-            // 
-            resources.ApplyResources(this.chkRtssOnlyActive, "chkRtssOnlyActive");
-            this.chkRtssOnlyActive.ForeColor = System.Drawing.Color.Black;
-            this.chkRtssOnlyActive.Name = "chkRtssOnlyActive";
-            this.chkRtssOnlyActive.UseVisualStyleBackColor = true;
-            // 
-            // chkStunEnable
-            // 
-            resources.ApplyResources(this.chkStunEnable, "chkStunEnable");
-            this.chkStunEnable.ForeColor = System.Drawing.Color.Black;
-            this.chkStunEnable.Name = "chkStunEnable";
-            this.chkStunEnable.UseVisualStyleBackColor = true;
             // 
             // label8
             // 
@@ -472,17 +400,8 @@ namespace tickMeter.Forms
             resources.ApplyResources(this, "$this");
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
             this.Controls.Add(this.btnAdvancedSettings);
-            this.Controls.Add(this.chkPingBindToInterface);
-            this.Controls.Add(this.chkPingTcpPrefer);
             this.Controls.Add(this.chkPingFallbackIcmp);
-            this.Controls.Add(this.chkPingTargetActiveOnly);
-            this.Controls.Add(this.chkTickrateSmoothing);
-            this.Controls.Add(this.chkDedupMultiNic);
-            this.Controls.Add(this.chkEnableIPv6);
             this.Controls.Add(this.chkIgnoreVirtualAdapters);
-            this.Controls.Add(this.chkRtssOnlyActive);
-            this.Controls.Add(this.chkStunEnable);
-            this.Controls.Add(this.captureAllAdaptersCheckbox);
             this.Controls.Add(this.local_ip_textbox);
             this.Controls.Add(this.local_ip_lbl);
             this.Controls.Add(this.run_minimized);
@@ -549,18 +468,8 @@ namespace tickMeter.Forms
         public TextBox local_ip_textbox;
         public ColorDialog colorDialog1;
         public CheckBox packet_drops_checkbox;
-        // NEW: Advanced universality checkboxes fields
-        public CheckBox chkPingBindToInterface;
-        public CheckBox chkPingTcpPrefer;
         public CheckBox chkPingFallbackIcmp;
-        public CheckBox chkPingTargetActiveOnly;
-        public CheckBox chkTickrateSmoothing;
-        public CheckBox chkDedupMultiNic;
-        public CheckBox chkEnableIPv6;
         public CheckBox chkIgnoreVirtualAdapters;
-        public CheckBox chkRtssOnlyActive;
-        public CheckBox chkStunEnable;
-        public CheckBox captureAllAdaptersCheckbox;
         public Label label8;
         public Label donate_lbl;
         public Label updateLbl;